grecka.c: do explicit convertions in ndm_time_add_sec() calls
grecka.c: In function ‘grecka_event_loop’:
grecka.c:591:26: warning: conversion to ‘int64_t’ {aka ‘const long int’}
from ‘long unsigned int’ may change the sign of the result [-Wsign-conversion]
  591 |    ndm_time_add_sec(&ts, interval);
      |                          ^~~~~~~~
grecka.c:598:39: warning: conversion to ‘int64_t’ {aka ‘const long int’}
from ‘long unsigned int’ may change the sign of the result [-Wsign-conversion]
  598 |    ndm_time_add_sec(&ts_rcv, interval * count);
      |                              ~~~~~~~~~^~~~~~~

+++ grecka.c
@@ -588,14 +588,14 @@ static void grecka_event_loop()
 			struct timespec ts_now;
 
 			ndm_time_get_monotonic(&ts_now);
-			ndm_time_add_sec(&ts, interval);
+			ndm_time_add_sec(&ts, (int64_t)interval);
 
 			if (ndm_time_greater_or_equal(&ts_now, &ts)) {
 				grecka_send_keepalive();
 				last_send = ts_now;
 			}
 
-			ndm_time_add_sec(&ts_rcv, interval * count);
+			ndm_time_add_sec(&ts_rcv, (int64_t)(interval * count));
 
 			if (ndm_time_greater_or_equal(&ts_now, &ts_rcv)) {
 				const char *args[] = {
